Dropdown was rendered too much (see commit message detail)
dropdown_ajax was rendering the dropdown every time it had
unacknowledged notifications. So if the user didn't click to trigger
call of dropdown_open, the dropdown would be rendered again.

To circunvent this, client side unacknowledged counts are passed to
dropdown_ajax which now compares the count of unacknowledged
notifications it has and the count of unacknowledged notification that
are already in the dropdown client side.

So dropdown_ajax started to fail because all the dropdowns declared in
the url weren't in the html and thus the counts where not passed. The
best fix for that was to add the missing declared queue in the template.